+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### 0.0.3 (SNAPSHOT)
-*
+* Handle process uncaught exception, log error instead of exit
 
 ### 0.0.2
 * Add custom express app config / set up hook